feat: Add raning on coments
